#8 change CSS spacing and move components files into folders
