+++ src/Components/ArticleLarge/ArticleLarge.css
@@ -38,6 +38,6 @@
 
 @media screen and (min-width: 992px) {
   .jumbotron {
-    height: 500px;
+    height: 380px;
   }
 }
