+++ src/Components/MainHeader/MainHeader.css
@@ -1,5 +1,5 @@
 .main-header {
-  margin: 88px auto 40px;
+  margin: 40px auto;
 }
   .main-header > * {
     margin-bottom: 0;
