+++ src/Components/TopBar/TopBar.css
@@ -8,6 +8,7 @@
   box-shadow: inset 0 -1px 0 0 #dddddd;
   background-color: #ffffff;
   color: rgb(111, 111, 111);
+  z-index: 100;
 }
 
 .top-bar * {
